Reworked based on review comments
