--- spring-jcl/src/main/java/org/apache/commons/logging/LogFactoryImpl.java
@@ -1,69 +1,0 @@
-/*
- * Copyright 2002-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.commons.logging;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-/**
- * A minimal implementation of the {@code LogFactory} class, delegating the
- * {@code Log} creation to {@code LogAdapter}.
- */
-class LogFactoryImpl extends LogFactory {
-
-	private final Map<String, Object> attributes = new ConcurrentHashMap<>();
-
-	@Override
-	public Log getInstance(Class<?> clazz) {
-		return getInstance(clazz.getName());
-	}
-
-	@Override
-	public Log getInstance(String name) {
-		return LogAdapter.createLog(name);
-	}
-
-	@Override
-	public void setAttribute(String name, Object value) {
-		if (value != null) {
-			this.attributes.put(name, value);
-		}
-		else {
-			this.attributes.remove(name);
-		}
-	}
-
-	@Override
-	public void removeAttribute(String name) {
-		this.attributes.remove(name);
-	}
-
-	@Override
-	public Object getAttribute(String name) {
-		return this.attributes.get(name);
-	}
-
-	@Override
-	public String[] getAttributeNames() {
-		return this.attributes.keySet().toArray(new String[0]);
-	}
-
-	@Override
-	public void release() {
-	}
-
-}
